Updating ReadMe and warranty2.sh
Adding iOS fields to output. Adding more verbose logging (when
enabled).

+++ warranty2.sh
@@ -13,6 +13,10 @@
 # SPX output format added by n8felton (02/27/2012)
 #
 # Last Edited 02/27/2012
+# Last Edited 2012/05/11
+# Adding iPhone Support
+# Fixing debugg and verbose flags
+# 
 
 ###############
 ##  GLOBALS  ##
@@ -28,8 +32,8 @@ PlistOutput="warranty.plist"
 SPXOutput="warranty.spx"
 Format="stdout"
 Version="5"
-NODEBUGG=	# Set to 1 to enable debugging ( Don't delete temp files )
-VERBOSE=	# Set to 1 to enable verbose l
+DEBUGG=		# Set to 1 to enable debugging ( Don't delete temp files )
+VERBOSE=	# Set to 1 to enable bulk editing verboseness
 
 #################
 ##  FUNCTIONS  ##
@@ -50,7 +54,9 @@ Output:
 	-f [csv|plist|spx|DSProperties] = FORMAT output file to csv, plist, spx, or DeployStudio format.
 	-o [/path/to/] = OUTPUT. Use Full Path. Do not include filename. Default is same directory as script.
 	-n <filename>[.plist|.csv|.spx] = Speficiy output filename. Ensure you use the appropriate extension for your output.
-	
+	-v = Enable sexy verboseness	
+	-k = Enable debugging (Don't delete temp files)
+
 Defaults:
 	WarrantyTempFile="${WarrantyTempFile}"
 	AsdCheck="${AsdCheck}"
@@ -83,6 +89,34 @@ Examples:
 	/usr/sbin/system_profiler -xml > firstreport.spx
 	${PlistBuddy} -c "Merge warranty.spx" firstreport.spx
 
+
+EOF
+}
+
+flags()
+{
+cat<<EOF
+
+Input:
+	no flags = use this computers serial
+	-b = loop through BULK csv file
+		 processing a csv file will only output to a 
+		 csv file. other output formats will be ignored
+	-s = specify SERIAL number
+
+Output:
+	-f [csv|plist|spx|DSProperties] = FORMAT output file to csv, plist, spx, or DeployStudio format.
+	-o [/path/to/] = OUTPUT. Use Full Path. Do not include filename. Default is same directory as script.
+	-n <filename>[.plist|.csv|.spx] = Speficiy output filename. Ensure you use the appropriate extension for your output.
+	-v = Enable sexy verboseness	
+	-k = Enable debugging (Don't delete temp files)
+	-h = Full help page (With Examples!)
+
+Defaults:
+	WarrantyTempFile="${WarrantyTempFile}"
+	AsdCheck="${AsdCheck}"
+	Output="${Output}"
+	Format="${Format}"
 EOF
 }
 
@@ -121,14 +155,24 @@ FixDate()
 outputPlist() {
 	PlistLocal="${Output}/${PlistOutput}"
 	# Create plist for output
+	if [ "${VERBOSE}" ]; then 
+		echo "Removing ${PlistLocal}"
+	fi
+	
 	rm "${PlistLocal}" > /dev/null 2>&1 # Probably Unnecessary, Just being Safe
+	
 	if [[ ! -e "${PlistLocal}" ]]; then
 		AddPlistString warrantyscriptversion "${Version}" "${PlistLocal}" > /dev/null 2>&1
-		for i in purchasedate warrantyexpires warrantystatus modeltype asd serialnumber currentdate daysremaining dayssincedop
+		for i in purchasedate warrantyexpires warrantystatus modeltype asd serialnumber currentdate daysremaining dayssincedop isaniphone iphonecarrier partdescript
 		do
 		AddPlistString $i unknown "${PlistLocal}"
 		done
 	fi
+	
+	if [ "${VERBOSE}" ]; then 
+		echo "Created ${PlistLocal}, adding fields"
+	fi
+	
 	SetPlistString serialnumber "${SerialNumber}" "${PlistLocal}"
 	SetPlistString purchasedate "${PurchaseDate}" "${PlistLocal}"
 	SetPlistString warrantyexpires "${WarrantyExpires}" "${PlistLocal}"
@@ -138,13 +182,24 @@ outputPlist() {
 	SetPlistString daysremaining "${DaysRemaining}" "${PlistLocal}"
 	SetPlistString dayssincedop "${DaysSinceDOP}" "${PlistLocal}"
 	SetPlistString currentdate $(date "+%m/%d/%Y") "${PlistLocal}"
+	SetPlistString isaniphone "${IsAniPhone}" "${PlistLocal}"
+	SetPlistString iphonecarrier "${iPhoneCarrier}" "${PlistLocal}"
+	SetPlistString partdescript "${PartDescript}" "${PlistLocal}"
+	
+	if [ "${VERBOSE}" ]; then 
+		echo "All fields added"
+	fi
 }
 
 outputCSV() {
-	# Csv output
+	
+	if [ "${VERBOSE}" ]; then 
+		echo "Creating csv output"
+	fi
 	# Serial#, PurchaseDate, DaysSinceDOP, WarrantyExpires, DaysRemaining, WarrantyStatus, ModelType, AsdVers
 	FixModel=$(echo ${ModelType} |tr -d ',')
-	echo "${SerialNumber}, ${PurchaseDate}, ${DaysSinceDOP}, ${WarrantyExpires}, ${DaysRemaining}, ${WarrantyStatus}, ${FixModel}, ${AsdVers}" >> "${Output}/${CSVOutput}"
+	echo "${SerialNumber}, ${PurchaseDate}, ${DaysSinceDOP}, ${WarrantyExpires}, ${DaysRemaining}, ${WarrantyStatus}, ${FixModel}, ${AsdVers}, ${IsAniPhone}, ${iPhoneCarrier}, ${PartDescript}" >> "${Output}/${CSVOutput}"
+	
 }
 
 outputSTDOUT() {
@@ -158,6 +213,9 @@ outputSTDOUT() {
 	echo "WarrantyStatus      ==  ${WarrantyStatus}"
 	echo "ModelType           ==  ${ModelType}"
 	echo "ASD                 ==  ${AsdVers}"
+	echo "IsAniPhone          ==  ${IsAniPhone}"
+	echo "iPhoneCarrier       ==  ${iPhoneCarrier}"
+	echo "PartDescript        ==  ${PartDescript}"
 }
 
 outputDSProperties() {
@@ -171,12 +229,20 @@ outputDSProperties() {
 	echo "RuntimeSetCustomProperty: WARRANTY_STATUS=${WarrantyStatus}"
 	echo "RuntimeSetCustomProperty: MODEL_TYPE=${ModelType}"
 	echo "RuntimeSetCustomProperty: ASD=${AsdVers}"
+	echo "RuntimeSetCustomProperty: IsAniPhone=${IsAniPhone}"
+	echo "RuntimeSetCustomProperty: iPhoneCarrier=${iPhoneCarrier}"
+	echo "RuntimeSetCustomProperty: PartDescript=${PartDescript}"
 	#A timestamp of the last time the information was polled for the "Days" information
 	echo "RuntimeSetCustomProperty: LAST_POLL=$(date +%Y-%m-%d)"	
 }
 
 outputSPX() {
+
 	SPXOutput=${Output}/${SPXOutput}
+	if [ "${VERBOSE}" ]; then 
+		echo "Outputing to SPX format."
+	fi
+
 (	
 cat <<EOF
 <?xml version="1.0" encoding="UTF-8"?>
@@ -188,6 +254,10 @@ cat <<EOF
 EOF
 ) > ${SPXOutput}
 
+	if [ "${VERBOSE}" ]; then 
+		echo "Adding Fields to SPX file."
+	fi
+
 	${PlistBuddy} -c "Add 0:_dataType string Warranty" ${SPXOutput}
 	${PlistBuddy} -c "Add 0:_parentDataType string SPHardwareDataType" ${SPXOutput}	
 	${PlistBuddy} -c "Add 0:_items array" ${SPXOutput}
@@ -198,25 +268,38 @@ EOF
 	${PlistBuddy} -c "Add 0:_items:0:Warranty\ Expires string ${WarrantyExpires}" ${SPXOutput}
 	${PlistBuddy} -c "Add 0:_items:0:Warranty\ Status string ${WarrantyStatus}" ${SPXOutput}
 	${PlistBuddy} -c "Add 0:_items:0:ASD string ${AsdVers}" ${SPXOutput}
+	${PlistBuddy} -c "Add 0:_items:0:IsAniPhone string ${IsAniPhone}" ${SPXOutput}
+	${PlistBuddy} -c "Add 0:_items:0:iPhoneCarrier string ${iPhoneCarrier}" ${SPXOutput}
+	${PlistBuddy} -c "Add 0:_items:0:PartDescript string ${PartDescript}" ${SPXOutput}	
 	${PlistBuddy} -c "Add 0:_properties:Model:_order string 1" ${SPXOutput}
 	${PlistBuddy} -c "Add 0:_properties:Serial\ Number:_order string 2" ${SPXOutput}
 	${PlistBuddy} -c "Add 0:_properties:Purchase\ Date:_order string 3" ${SPXOutput}
 	${PlistBuddy} -c "Add 0:_properties:Warranty\ Expires:_order string 4" ${SPXOutput}
 	${PlistBuddy} -c "Add 0:_properties:Warranty\ Status:_order string 5" ${SPXOutput}
 	${PlistBuddy} -c "Add 0:_properties:ASD:_order string 6" ${SPXOutput}
+	${PlistBuddy} -c "Add 0:_properties:IsAniPhone:_order string 7" ${SPXOutput}
+	${PlistBuddy} -c "Add 0:_properties:iPhoneCarrier:_order string 8" ${SPXOutput}
+	${PlistBuddy} -c "Add 0:_properties:PartDescript:_order string 9" ${SPXOutput}
+
+	if [ "${VERBOSE}" ]; then 
+		echo "Added All Fields to SPX file."
+	fi	
 }
 
 processCSV() {
-if [ "$VERBOSE" ]; then 
+if [ "${VERBOSE}" ]; then 
 	echo "Creating ${Output}/${CSVOutput}"
 fi
 
 # Echo headings into CSV file.
-echo "SerialNumber, PurchaseDate, DaysSinceDOP, WarrantyExpires, DaysRemaining, WarrantyStatus, FixModel, AsdVers" >> "${Output}/${CSVOutput}"
+if [ "${VERBOSE}" ]; then 
+	echo "Adding CSV headers."
+fi
+echo "SerialNumber, PurchaseDate, DaysSinceDOP, WarrantyExpires, DaysRemaining, WarrantyStatus, FixModel, AsdVers, IsAniPhone, iPhoneCarrier, PartDescript" >> "${Output}/${CSVOutput}"
 
 for i in $(cat "${1}"); do
 
-if [ "$VERBOSE" ]; then 
+if [ "${VERBOSE}" ]; then 
 	echo "Checking ${i}"
 fi
 
@@ -225,13 +308,16 @@ checkStatus
 outputCSV
 sleep 5
 done
-exit 0
 
 }
 
 checkStatus() {
 
 WarrantyURL="https://selfsolve.apple.com/warrantyChecker.do?sn=${SerialNumber}&country=USA"
+
+if [ "${VERBOSE}" ]; then 
+	echo "Checking Serial: ${SerialNumber} Warranty Status from URL ${WarrantyURL}"
+fi
 
 [[ -n "${SerialNumber}" ]] && WarrantyInfo=$(curl -k -s $WarrantyURL | awk '{gsub(/\",\"/,"\n");print}' | awk '{gsub(/\":\"/,":");print}' | sed s/\"\}\)// > ${WarrantyTempFile})
 
@@ -241,16 +327,24 @@ InvalidSerial=$(grep 'invalidserialnumber\|productdoesnotexist' "${WarrantyTempF
 
 if [[ -n "${InvalidSerial}" ]]; then
 	WarrantyStatus="Serial Invalid: ${SerialNumber}. ${InvalidSerial}"
+	if [ "${VERBOSE}" ]; then 
+		echo "Invalid Serial Number"
+	fi
 fi
 	
 if [[ -z "${SerialNumber}" ]]; then 
 	WarrantyStatus="Serial Missing: ${SerialNumber}. ${InvalidSerial}"
+	if [ "${VERBOSE}" ]; then 
+		echo "Serial Number Missing"
+	fi
 fi
-
 
 
 if [[ -e "${WarrantyTempFile}" && -z "${InvalidSerial}" ]] ; then
 
+	if [ "${VERBOSE}" ]; then 
+		echo "Scanning file for specified fields."
+	fi
 	PurchaseDate=$(GetWarrantyValue PURCHASE_DATE)
 	WarrantyStatus=$(GetWarrantyValue HW_SUPPORT_COV_SHORT)
 	WarrantyExpires=$(GetWarrantyValue HW_END_DATE)
@@ -274,9 +368,24 @@ if [[ -e "${WarrantyTempFile}" && -z "${InvalidSerial}" ]] ; then
 	#Days remaining
 	DaysRemaining=$(GetWarrantyValue DAYS_REM_IN_COV)
 	
+	#Is serial an iPhone?
+	IsAniPhone=$(GetWarrantyValue IS_IPHONE)
+	
+	#iPhone Carrier?
+	iPhoneCarrier=$(GetWarrantyValue CARRIER)
+	
+	#iPhone Part Description
+	PartDescript=$(GetWarrantyValue PART_DESCR)
+	
+	if [ "${VERBOSE}" ]; then 
+		echo "All Fields Found."
+	fi
 	# Delete temp files if NODEBUGG is 1
-	if [[ "${NODEBUGG}" ]]; then
+	if [ ! "${DEBUGG}" ]; then
 		rm "${WarrantyTempFile}"
+		if [ "${VERBOSE}" ]; then 
+			echo "Removing Temp Warranty File"
+		fi
 	fi
 fi
 
@@ -286,6 +395,10 @@ PrintData()
 {
 case $Format in
 	csv)
+	if [ "${VERBOSE}" ]; then 
+	echo "Adding CSV Headers."
+	fi
+	echo "SerialNumber, PurchaseDate, DaysSinceDOP, WarrantyExpires, DaysRemaining, WarrantyStatus, FixModel, AsdVers, IsAniPhone, iPhoneCarrier, PartDescript" >> "${Output}/${CSVOutput}"
 	outputCSV
 	;;
 	plist)
@@ -316,7 +429,7 @@ FixDateYYYYMMDD()
 # Get output options: csv, plist
 # You should be exporting this to a static place you can pull reports from.
 
-while getopts s:b:o:f:n:dvh opt; do
+while getopts s:b:o:f:n:dvkh opt; do
 	case "$opt" in
 		s) SerialNumber="$OPTARG";;
 		b) SerialCSV="$OPTARG";;
@@ -325,10 +438,11 @@ while getopts s:b:o:f:n:dvh opt; do
 		n) OutputName="$OPTARG";;
 		d) FixDate=1;;
 		v) VERBOSE=1;;
+		k) DEBUGG=1;;
 		h) help
 			exit 0;;
 		\?)
-			help
+			flags
 			exit 0;;
 	esac
 done
@@ -337,27 +451,45 @@ shift $(expr $OPTIND - 1)
 
 ## Add Timeout so this curl doesn't try forever when GitHub raw is down. Cache a local copy? Try alternate forks?
 ## Fail to unknown ASD after 5 seconds? Don't take too long, you'll get the numbers eventually. 
-
+if [ "${VERBOSE}" ]; then 
+	echo "Downloading ASD file."
+fi
 curl -k -s https://raw.github.com/rustymyers/warranty/master/asdcheck -o ${AsdCheck} #> /dev/null 2>&1
 
 # No command line variables. Use internal serial and run checks
 if [[ -z "$SerialNumber" ]]; then
+	if [ "${VERBOSE}" ]; then 
+		echo "No Serial's provided, using local computers"
+	fi
 	SerialNumber=$(system_profiler SPHardwareDataType | grep "Serial Number" | grep -v "tray" |  awk -F ': ' {'print $2'} 2>/dev/null)
 fi
 
 if [ "${OutputName}" ]; then
+	if [ "${VERBOSE}" ]; then 
+		echo "Changing output name to ${OutputName}"
+	fi
 	CSVOutput="${OutputName}"
 	PlistOutput="${OutputName}"
 	SPXOutput="${OutputName}"
 fi
 
 if [[ -z "$SerialCSV" ]]; then
+
 checkStatus
 PrintData
 else
+if [ "${VERBOSE}" ]; then 
+	echo "Processing CSV file."
+fi
 processCSV "${SerialCSV}"
 fi
 
-rm "${AsdCheck}"
+# Delete temp files if NODEBUGG is 1
+if [ ! "${DEBUGG}" ]; then
+	rm "${AsdCheck}"
+		if [ "${VERBOSE}" ]; then 
+			echo "Removing Temp ASD File"
+		fi
+fi
 
 exit 0
